Crate seq_arithmetic

Seq Arithmetic

This library contains several handy functions to do basic arithmetic on byte sequences

There are several different ways of creating byte sequences, for example by converting from other numeric types.

If you wish to create a byte sequence directly, the easiest way we have found is to do the following:

bytes!(BSeq, 2);
 
let mySeq = BSeq(secret_bytes!([0x0fu8, 0x15u8])).to_be_bytes();

Note that you have to specify the number of bytes you wish to input. When inputting multiple byte sequences of different length, you can avoid defining multiple types by defining a large sequence, padding with leading zeroes, and use be_seq_trim() to trim the zeroes away afterwards.

bytes!(BSeq, 6);
 
let myLongSeq = BSeq(secret_bytes!([
    0x0fu8, 0x15u8, 
    oxffu8, 0x0fu8, 
    0x15u8, oxffu8
])).to_be_bytes();
 
let myShortSeq = be_seq_trim(&BSeq(secret_bytes!([
    0x00u8, 0x00u8, 
    ox00u8, 0x00u8, 
    0x15u8, oxffu8
])).to_be_bytes());

Functions

be_U32_to_seq

Converts a U32 to Seq

be_seq_add

Adds two byte sequences of any length

be_seq_div

Makes integer division on two byte sequences of any length and returns (quotient, remainder) Returns 0 when trying to divide by 0 TODO: Give error when dividing by 0

be_seq_exp

Calculates exponents NOTE: only supports exponents of max size usize TODO: Add support of any size exponents, perhaps put loops in a seperate function that supports loops to a byteSeq max

be_seq_mod

Modular operation. Finds the rest when dividing two numbers

be_seq_mod_exp

Performs modular exponentiation i. e. a^b mod n Is designed to avoid extremely large intermediate values Uses the square multiply algorithm

be_seq_mul

Mulitplies two byte sequences of any length and returns the product

be_seq_mul_mod

Performs both multiplication and modular operation

be_seq_sub

Subtracts two byte sequences of any length, returns result as well as bool stating if theres underflow or not

be_seq_trim

Trims a byte sequence by removing leading zeroes. Returns the trimmed byte sequence

seq_eq

Trims and compares two byte sequences, returns true if equal

seq_leq

Returns true if a is less or equal to b

seq_one

returns byte sequence with value one

seq_shift_left

Adds n trailing bytes with value zero to a

seq_shift_right

Adds n leading bytes with value zero to a

seq_to_U128

Converts a byte sequence to a U128

seq_to_usize

Converts a byte sequence to usize

seq_zero

returns byte sequence with value zero